+++ ft_strchr.c
@@ -6,7 +6,7 @@
 /*   By: fboulbes <fboulbes@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/10/31 14:03:20 by fboulbes          #+#    #+#             */
-/*   Updated: 2024/10/31 14:06:54 by fboulbes         ###   ########.fr       */
+/*   Updated: 2024/11/05 13:55:27 by fboulbes         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 char	*ft_strchr(const char *string, int searchedChar)
 {
-	int	i;
+	size_t	i;
 
 	i = 0;
 	while (string[i] != (char)searchedChar)
